Добавила hook для добавления active на кнопку навигации при рендере страницы
